Refactor `_getVisitsCollection` into patient-visits mixin
